Consolidate common rhsso functionality
